Code cleanup - move RaptorRouter up to package org.opentripplanner.routing.algorithm.
